refactor(interpreter): split variables ranges and registration
